+++ main.go
@@ -1,12 +1,9 @@
 package main
 
 import (
-	"bufio"
 	"flag"
 	"fmt"
-	"io"
 	"os"
-	"regexp"
 
 	"github.com/koron/gomigemo/embedict"
 	"github.com/koron/gomigemo/migemo"
@@ -20,33 +17,6 @@ type grepOpt struct {
 	optNumber   bool
 	optFilename bool
 	filename    string
-}
-
-// Does the grepping
-func grep(r io.Reader, re *regexp.Regexp, opt *grepOpt) error {
-	buf := bufio.NewReader(r)
-	n := 1
-	for {
-		b, _, err := buf.ReadLine()
-		if err != nil {
-			if err == io.EOF {
-				return nil
-			}
-			return err
-		}
-		line := string(b)
-		if re.MatchString(line) {
-			if opt.optFilename {
-				fmt.Printf("%s:", opt.filename)
-			}
-			if opt.optNumber {
-				fmt.Printf("%d:", n)
-			}
-			fmt.Println(line)
-			n++
-		}
-	}
-	return nil
 }
 
 func main() {
